Implemented props synchronization
