--- bnb-cocos/Classes/BnbGameServer/BnbGameServer/Debug/BnbGameServer.log
@@ -1,3 +1,0 @@
-﻿  BnbGameServer.cpp
-  BnbGameServer.vcxproj -> c:\users\administrator\documents\visual studio 2017\Projects\BnbGameServer\Debug\BnbGameServer.exe
-  BnbGameServer.vcxproj -> c:\users\administrator\documents\visual studio 2017\Projects\BnbGameServer\Debug\BnbGameServer.pdb (Partial PDB)
