--- bnb-cocos/Classes/BnbGameServer/BnbGameServer/Debug/BnbGameServer.tlog/BnbGameServer.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.15063.0
-Debug|Win32|c:\users\administrator\documents\visual studio 2017\Projects\BnbGameServer\|
