--- bnb-cocos/Classes/BnbGameServer/BnbGameServer/main.cpp
@@ -1,8 +1,0 @@
-#include "BnbGameServer.h"
-#include <iostream>
-using std::cout;
-GameServer gc;
-int main()
-{
-	gc.ProcessGameServer();
-}
