+++ bnb-cocos/Classes/BnbGameServer/main.cpp
@@ -1,0 +1,8 @@
+#include "BnbGameServer.h"
+#include <iostream>
+using std::cout;
+GameServer gc;
+int main()
+{
+	gc.ProcessGameServer();
+}
