+++ bnb-cocos/Classes/WebGameScene.h
@@ -13,12 +13,14 @@ public:
 	static cocos2d::Scene* createScene();
 	virtual bool init();    
 	void update(float dt) override;
+	void acceptProps();
 	//继承的GameScene里面，已经创建了一个_myplayer，该chara在本client操作
 	//所以在这里添加的其他chara都是非本client的
-	
+	virtual ~WebGameScene();
 	CREATE_FUNC(WebGameScene);
 private:
 	GameClient client;
 	GameScene* runningGameScene;
 	character* mychara;
+	char *prop;
 };
